# docs/swagger.yaml
basePath: /api
definitions:
...
  models.CreateTeamRequest:
    properties:
      city:
        description: Город
        type: string
      description:
        description: Описание
        type: string
      logo:
        description: Логотип
        type: string
      media:
        description: Медиа
        type: string
      name:
        description: Название*
        type: string
      participant_count:
        description: Кол-во участников
        type: integer
      uniform_color:
        description: Цвет формы
        type: string
    required:
    - city
    - name
    type: object
...
  models.Media:
    properties:
      created_at:
        type: string
      ext:
        type: string
      id:
        type: integer
      name:
        type: string
      path:
        type: string
      size:
        type: integer
    type: object
...
  models.Role:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.TeamView:
    properties:
      city:
        description: Город
        type: string
      created_at:
        description: Дата создания
        type: string
      description:
        description: Описание
        type: string
      disability_category:
        description: Категория инвалидности
        type: string
      id:
        type: integer
      logo:
        allOf:
        - $ref: '#/definitions/models.Media'
        description: Логотип
      media:
        description: Медиа
        items:
          $ref: '#/definitions/models.Media'
        type: array
      name:
        description: Название
        type: string
      participant_count:
        description: Кол-во участников
        type: integer
      responsible:
        allOf:
        - $ref: '#/definitions/models.UserView'
        description: Ответственный
      status:
        description: Статус
        type: integer
      uniform_color:
        description: Цвет формы
        type: string
    type: object
...
  models.UpdateTeamRequest:
    properties:
      city:
        description: Город
        type: string
      description:
        description: Описание
        type: string
      disability_category:
        description: Категория инвалидности
        type: string
      logo:
        description: Логотип
        type: string
      media:
        description: Медиа
        type: string
      name:
        description: Название*
        type: string
      participant_count:
        description: Кол-во участников
        type: integer
      uniform_color:
        description: Цвет формы
        type: string
    required:
    - city
    - name
    type: object
...
  models.UserView:
    properties:
      city:
        description: Город
        type: string
      created_at:
        description: Дата создания
        type: string
      email:
        description: Email
        type: string
      id:
        type: integer
      logo:
        description: Логотип
        type: string
      media:
        description: Медиа
        type: string
      name:
        description: ФИО
        type: string
      phone:
        description: Телефон
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: Роль
      status:
        description: Статус
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Приложения бронирования спортивных площадок.
  title: Спортивный город
  version: "1.0"
paths:
...
  /api/teams:
    get:
      consumes:
      - application/json
      description: Получение списка всех команд
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TeamView'
            type: array
        "400":
          description: Bad Request
          schema:
            type: Bad
        "500":
          description: Internal Server Error
          schema:
            type: Internal
      summary: Возвращает список всех команд
      tags:
      - Команды
    post:
      description: Создание новой команды
      parameters:
      - description: Данные для создания новой команды
        in: body
        name: createTeam
        required: true
        schema:
          $ref: '#/definitions/models.CreateTeamRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TeamView'
        "422":
          description: Unprocessable Entity
          schema:
            type: Unprocessable
      summary: Создание новой команды
      tags:
      - Команды
  /api/teams/{id}:
    delete:
      description: Удаление команды по идентификатору
      parameters:
      - description: ID команды
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: "No"
        "404":
          description: Not Found
          schema:
            type: Not
      summary: Удаляет команду по ID
      tags:
      - Команды
    get:
      description: Получение информации о команде по идентификатору
      parameters:
      - description: ID команды
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TeamView'
        "400":
          description: Bad Request
          schema:
            type: Bad
        "404":
          description: Not Found
          schema:
            type: Not
      summary: Возвращает информацию о команде по ID
      tags:
      - Команды
    put:
      description: Обновление существующей команды
      parameters:
      - description: Данные для обновления команды
        in: body
        name: updateTeam
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTeamRequest'
      - description: ID команды
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: "No"
        "404":
          description: Not Found
          schema:
            type: Not
        "422":
          description: Unprocessable Entity
          schema:
            type: Unprocessable
      summary: Обновление существующей команды
      tags:
      - Команды
...
swagger: "2.0"